refactor!: fully migrate to app/
Deleted `pages/` dir

* delete unused `UnderConstruction` component

* fully update to new SEO configs

* update `CustomHead` and deprecate `next-seo`
  * no more `CustomHead2`

* update `ConfettiButton`
