+++ package.json
@@ -21,9 +21,7 @@
         "gray-matter": "^4.0.3",
         "next": "13.5.1",
         "next-mdx-remote": "^4.4.1",
-        "next-seo": "^6.0.0",
         "next-themes": "^0.2.1",
-        "nextjs-progressbar": "^0.0.16",
         "nprogress": "^0.2.0",
         "postcss": "8.4.31",
         "react": "18.2.0",
@@ -60,7 +58,6 @@
         "husky": "^8.0.0",
         "lint-staged": "^14.0.1",
         "next-sitemap": "^4.2.3",
-        "npm-run-all": "^4.1.5",
         "prettier": "^2.8.8",
         "prettier-plugin-pkg": "^0.17.1",
         "prettier-plugin-sh": "^0.12.8",
