+++ src/components/ConfettiButton.tsx
@@ -1,9 +1,13 @@
 'use client'
 
 import confetti from 'canvas-confetti'
-import Button, { type ButtonProps } from './Button'
+import Button from './Button'
 
-export default function ConfettiButton(props: ButtonProps) {
+export default function ConfettiButton({
+    children,
+}: {
+    children: React.ReactNode
+}) {
     const confettiOptions: (angle: number) => confetti.Options = angle => ({
         angle,
         spread: randomInRange(45, 80),
@@ -15,7 +19,11 @@ export default function ConfettiButton(props: ButtonProps) {
         confetti({ origin: { x: 1 }, ...confettiOptions(100) })
     }
 
-    return <Button {...props} onClick={celebrate} />
+    return (
+        <Button attractive onClick={celebrate}>
+            {children}
+        </Button>
+    )
 }
 
 const randomInRange = (min: number, max: number) =>
