+++ src/components/CustomHead.tsx
@@ -1,89 +1,24 @@
-import { DefaultSeo } from 'next-seo'
-import { useRouter } from 'next/router'
+'use client'
+
+import { usePathname } from 'next/navigation'
+import Script from 'next/script'
 import { MY_DOMAIN } from '@/utils/config'
-import { colors } from '@/utils/tailwind'
 
-const DEFAULT_TITLE = 'Abdullah Zeidan | Web Engineer'
-const DEFAULT_DESCRIPTION =
-    'A dev with the passion to make the web a greater place.'
-
+/** A custom `<head>` for when `generateMetadata()` doesn't suffice. */
 export default function CustomHead() {
-    const router = useRouter()
-    const canonicalURL = `${MY_DOMAIN}${router.asPath}`
+    const pathname = usePathname()
+    const CANONICAL_URL = `${MY_DOMAIN}${pathname}`
 
     return (
-        <DefaultSeo
-            title={DEFAULT_TITLE}
-            description={DEFAULT_DESCRIPTION}
-            canonical={canonicalURL}
-            twitter={{
-                site: '@zabd_dev',
-                cardType: 'summary_large_image',
-            }}
-            themeColor={colors.dark}
-            openGraph={{
-                title: DEFAULT_TITLE,
-                description: DEFAULT_DESCRIPTION,
-                url: canonicalURL,
-                type: 'website',
-                siteName: "Abdullah Zeidan's Personal Website/Portfolio",
-                images: [
-                    {
-                        url: `${MY_DOMAIN}/images/og-image.png`,
-                        secureUrl: `${MY_DOMAIN}/images/og-image.png`,
-                        type: 'image/png',
-                        width: 256,
-                        height: 256,
-                        alt: "Abdullah Zeidan's logo",
-                    },
-                ],
-            }}
-            additionalMetaTags={[
-                {
-                    name: 'google-site-verification',
-                    content: 'sF_y0oto8QSBuruHbL7znfMWdsgBa3Ljk9jFwIpZYEI',
-                },
-                {
-                    name: 'viewport',
-                    content: 'width=device-width, initial-scale=1',
-                },
-                {
-                    name: 'application-name',
-                    content: 'zabd.dev',
-                },
-                {
-                    name: 'msapplication-TileColor',
-                    content: colors.dark,
-                },
-            ]}
-            additionalLinkTags={[
-                {
-                    rel: 'apple-touch-icon',
-                    sizes: '180x180',
-                    href: '/images/apple-touch-icon.png',
-                },
-                {
-                    rel: 'icon',
-                    type: 'image/svg',
-                    sizes: '256x256',
-                    href: '/favicon.svg',
-                },
-                {
-                    rel: 'icon',
-                    type: 'image/png',
-                    sizes: '32x32',
-                    href: '/favicon-32x32.png',
-                },
-                {
-                    rel: 'manifest',
-                    href: '/meta/site.webmanifest',
-                },
-                {
-                    rel: 'mask-icon',
-                    href: '/images/safari-pinned-tab.svg',
-                    color: colors.brand[600],
-                },
-            ]}
-        />
+        // eslint-disable-next-line @next/next/no-head-element
+        <head>
+            <link rel='canonical' href={CANONICAL_URL} />
+            <meta property='og:url' content={CANONICAL_URL} />
+            <Script
+                async
+                src='https://analytics.zabd.dev/script.js'
+                data-website-id='a9d78ec7-c5cd-48c3-8b4c-43eb2ce3ad9d'
+            />
+        </head>
     )
 }
