--- src/components/Seo.tsx
@@ -1,7 +1,0 @@
-'use client'
-
-import { NextSeo, type NextSeoProps } from 'next-seo'
-
-export default function Seo(props: NextSeoProps) {
-    return <NextSeo {...props} />
-}
