+++ src/posts/welcome.mdx
@@ -6,7 +6,7 @@ isPublished: true
 
 ## Hey There! 👋
 
-<ConfettiButton attractive children='Celebrate!' />
+<ConfettiButton>Celebrate!</ConfettiButton>
 
 It looks like you're interested enough to take a look at my very first blog post.
 I would like to begin with welcoming you to the launch of my blog! Hopefully, it will be a place for many exciting stuff for me, and hopefully for you, too.
